coffeeAGNTCY/coffee_agents/recruiter/pyproject.toml
[project]
name = "agent-recruiter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk==0.3.2",
    "mcp>=1.24.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "litellm>=1.80.5",
    "langchain-litellm>=0.3.0",
    "langchain-anthropic>=0.3.13",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.16",
    "agntcy-app-sdk==0.4.6",
    "google-adk>=1.0.0",
    "rogue-ai-sdk>=0.3.4",
]

[project.optional-dependencies]
dev = [
    "agntcy-oasf-sdk-grpc-python==1.75.0.1.20250917120021+8b2bf93bf8dc",
    "agntcy-oasf-sdk-protocolbuffers-python==32.1.0.1.20250917120021+8b2bf93bf8dc",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_recruiter"]

[tool.pytest.ini_options]
testpaths = ["test"]
addopts = "--asyncio-mode=auto"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
